tuiEdge now stable enough
